move debug test into _testdata
