+++ cl/_testdata/debug/flags.txt
@@ -1,0 +1,1 @@
+-dbg
