+++ cl/_testdata/debug/out.ll
@@ -1,0 +1,1 @@
+;
