+++ cl/cltest/cltest.go
@@ -159,7 +159,7 @@ func TestCompileEx(t *testing.T, src any, fname, expected string) {
 	pkg := types.NewPackage(name, name)
 	imp := packages.NewImporter(fset)
 	foo, _, err := ssautil.BuildPackage(
-		&types.Config{Importer: imp}, fset, pkg, files, ssa.SanityCheckFunctions|ssa.InstantiateGenerics)
+		&types.Config{Importer: imp}, fset, pkg, files, ssa.SanityCheckFunctions|ssa.InstantiateGenerics|ssa.GlobalDebug)
 	if err != nil {
 		t.Fatal("BuildPackage failed:", err)
 	}
